+++ vgm.h
@@ -113,17 +113,17 @@ typedef struct vgm_s
 	char *creator;          // VGM creator name
 	char *notes;            // notes
     // Playback control
-    uint32_t sample_rate;       // sample rate
-    uint32_t data_pos;          // position of current data
-    uint32_t samples_waiting;   // # of samples waiting, in 44100Hz unit
-    nesapu_t *apu;              // NES APU
+    unsigned int sample_rate;       // sample rate
+    size_t data_pos;                // position of current data
+    unsigned int samples_waiting;   // # of samples waiting, in 44100Hz unit
+    nesapu_t *apu;                  // NES APU
  } vgm_t;
 
 
 vgm_t* vgm_create(file_reader_t *reader);
 void vgm_destroy(vgm_t *vgm);
-bool vgm_prepare_playback(vgm_t *vgm, uint32_t srate);
-int vgm_get_sample(vgm_t *vgm, int16_t *buf, int size);
+bool vgm_prepare_playback(vgm_t *vgm, unsigned int srate);
+int vgm_get_sample(vgm_t *vgm, int16_t *buf, unsigned int size);
 
 #ifdef __cplusplus
 }
